+++ src/gettsim_personas/__init__.py
@@ -1,10 +1,6 @@
 """Personas to use as example data when calling GETTSIM."""
 
-from pathlib import Path
-
-from gettsim_personas.active_personas import get_personas, upsert_input_data
+from gettsim_personas.active_personas import get_personas
+from gettsim_personas.upsert import upsert_input_data
 
 __all__ = ["get_personas", "upsert_input_data"]
-
-
-PERSONAS_SOURCE_DIR = Path(__file__).parent / "personas"
