+++ src/gettsim_personas/upsert.py
@@ -1,5 +1,3 @@
-"""Functions to broadcast GETTSIM input data."""
-
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
@@ -9,7 +7,69 @@ import numpy as np
 import pandas as pd
 
 if TYPE_CHECKING:
-    from gettsim_personas.active_personas import NestedDataDict
+    from gettsim_personas.typing import NestedDataDict
+
+
+def upsert_input_data(
+    input_data: NestedDataDict,
+    data_to_upsert: NestedDataDict,
+) -> NestedDataDict:
+    """Upsert GETTSIM input data.
+
+    Updates and inserts new values into a NestedDataDict with input data for GETTSIM,
+    e.g. created from a persona via `get_personas`. Data not in `data_to_upsert` is
+    broadcasted to match the length of `data_to_upsert`.
+
+    The length of data in `data_to_upsert` must be a multiple of the length of data in
+    `input_data`.
+
+    Args:
+        input_data:
+            NestedDataDict with input data for GETTSIM. Typically created from a persona
+            object via `get_personas`.
+        data_to_upsert:
+            NestedDataDict with data to be upserted
+
+    Returns:
+        NestedDataDict with upserted data
+    """
+    _fail_if_data_to_upsert_is_not_dict_with_array_or_series_leafs(data_to_upsert)
+    _fail_if_data_lengths_are_incompatible(data_to_upsert, input_data)
+    flat_data_to_upsert = dt.flatten_to_tree_paths(data_to_upsert)
+    flat_input_data = dt.flatten_to_tree_paths(input_data)
+
+    expected_length = len(next(iter(flat_data_to_upsert.values())))
+    persona_length = len(next(iter(flat_input_data.values())))
+    number_of_new_personas = expected_length // persona_length
+
+    upserted_data = flat_data_to_upsert.copy()
+    for path, series in flat_input_data.items():
+        if path in upserted_data:
+            continue
+        if path == ("p_id",):
+            broadcasted_series = broadcast_p_id(
+                original_series=series,
+                expected_length=expected_length,
+            )
+        elif "p_id_" in path[-1]:
+            broadcasted_series = broadcast_foreign_keys(
+                original_series=series,
+                expected_length=expected_length,
+            )
+        elif path[-1].endswith("_id"):
+            broadcasted_series = broadcast_group_ids(
+                original_series=series, expected_length=expected_length
+            )
+        else:
+            broadcasted_series = pd.Series(
+                np.tile(series.values, number_of_new_personas)
+            )
+
+        broadcasted_series.name = path[-1]
+        broadcasted_series.index = pd.RangeIndex(expected_length)
+        upserted_data[path] = broadcasted_series
+
+    return dt.unflatten_from_tree_paths(upserted_data)
 
 
 def broadcast_p_id(original_series: pd.Series, expected_length: int) -> pd.Series:
